backend/requirements.txt
fastapi
uvicorn
cohere
qdrant-client
psycopg2-binary
python-dotenv
pydantic
langchain-text-splitters
